workflows/pyproject.toml
[project]
name = "workflows"
version = "0.0.0"
description = "Hackathon Paris"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mistralai[realtime]>=1.12.4",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "requests>=2.32.3",
    "temporalio>=1.23.0",
]
